chg: vector ref to ptr
