+++ src/filters/sequencer/name_maps.h
@@ -7,7 +7,7 @@
 
 namespace dmxfish::filters {
     struct name_maps {
-        std::unordered_map<std::string, size_t> name_to_id_8bit, name_to_id_16bit, name_to_id_float, name_to_id_color;
+        std::map<std::string, size_t> name_to_id_8bit, name_to_id_16bit, name_to_id_float, name_to_id_color;
         name_maps() : name_to_id_8bit(), name_to_id_16bit(), name_to_id_float(), name_to_id_color() {}
     };
 }
